routes pour l'authentification validée

+++ routes/api.php
@@ -1,6 +1,7 @@
 <?php
 
 use Illuminate\Http\Request;
+use App\Http\Controllers\Auth\SocialAuthController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\BlogController;
 use App\Http\Controllers\BlogCommentController;
@@ -23,16 +24,45 @@ Route::get('/test', function () {
     return response()->json(['message' => 'API works!']);
 });
 
+
+Route::prefix('v1')->group(function () {
+    Route::post('/register', [AuthController::class, 'register']);
+    Route::post('/login', [AuthController::class, 'login']);
+    Route::post('/social-login', [AuthController::class, 'socialLogin']);
+
+    Route::middleware('auth:api')->group(function () {
+        Route::post('/logout', [AuthController::class, 'logout']);
+        Route::get('/profile', function (Request $request) {
+            return $request->user()->makeHidden(['password']);
+        });
+    });
+
+
+     Route::post('/password/email', [AuthController::class, 'sendPasswordResetLink']);
+    Route::post('/password/reset', [AuthController::class, 'resetPassword']);
+
+    Route::get('/email/verify/{id}/{hash}', [AuthController::class, 'verifyEmail'])
+        ->middleware('signed')
+        ->name('verification.verify');
+
+    Route::post('/2fa/enable', [AuthController::class, 'enable2FA'])->middleware('auth:api');
+
+    Route::get('auth/{provider}', [SocialAuthController::class, 'redirect']);
+    Route::get('auth/{provider}/callback', [SocialAuthController::class, 'callback']);
+});
+
+
+
 Route::prefix('v1')->group(function () {
     // Blogs
     Route::prefix('blogs')->group(function () {
         Route::get('/', [BlogController::class, 'index']);
         Route::get('{blog}', [BlogController::class, 'show']);
         Route::get('slug/{blog:slug}', [BlogController::class, 'show']);
-        Route::post('/', [BlogController::class, 'store']);
-        Route::put('{blog}', [BlogController::class, 'update']);
+        Route::post('/', [BlogController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{blog}', [BlogController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{blog:slug}', [BlogController::class, 'update']);
-        Route::delete('{blog}', [BlogController::class, 'destroy']);
+        Route::delete('{blog}', [BlogController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{blog:slug}', [BlogController::class, 'destroy']);
     });
 
@@ -62,10 +92,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [FaqController::class, 'index']);
         Route::get('{faq}', [FaqController::class, 'show']);
         Route::get('slug/{faq:slug}', [FaqController::class, 'show']);
-        Route::post('/', [FaqController::class, 'store']);
-        Route::put('{faq}', [FaqController::class, 'update']);
+        Route::post('/', [FaqController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{faq}', [FaqController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{faq:slug}', [FaqController::class, 'update']);
-        Route::delete('{faq}', [FaqController::class, 'destroy']);
+        Route::delete('{faq}', [FaqController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{faq:slug}', [FaqController::class, 'destroy']);
     });
 
@@ -86,10 +116,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [ProductController::class, 'index']);
         Route::get('{product}', [ProductController::class, 'show']);
         Route::get('slug/{product:slug}', [ProductController::class, 'show']);
-        Route::post('/', [ProductController::class, 'store']);
-        Route::put('{product}', [ProductController::class, 'update']);
+        Route::post('/', [ProductController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{product}', [ProductController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{product:slug}', [ProductController::class, 'update']);
-        Route::delete('{product}', [ProductController::class, 'destroy']);
+        Route::delete('{product}', [ProductController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{product:slug}', [ProductController::class, 'destroy']);
     });
 
@@ -98,10 +128,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [TestimonialController::class, 'index']);
         Route::get('{testimonial}', [TestimonialController::class, 'show']);
         Route::get('slug/{testimonial:slug}', [TestimonialController::class, 'show']);
-        Route::post('/', [TestimonialController::class, 'store']);
-        Route::put('{testimonial}', [TestimonialController::class, 'update']);
+        Route::post('/', [TestimonialController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{testimonial}', [TestimonialController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{testimonial:slug}', [TestimonialController::class, 'update']);
-        Route::delete('{testimonial}', [TestimonialController::class, 'destroy']);
+        Route::delete('{testimonial}', [TestimonialController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{testimonial:slug}', [TestimonialController::class, 'destroy']);
     });
 
@@ -110,10 +140,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [WebinarController::class, 'index']);
         Route::get('{webinar}', [WebinarController::class, 'show']);
         Route::get('slug/{webinar:slug}', [WebinarController::class, 'show']);
-        Route::post('/', [WebinarController::class, 'store']);
-        Route::put('{webinar}', [WebinarController::class, 'update']);
+        Route::post('/', [WebinarController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{webinar}', [WebinarController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{webinar:slug}', [WebinarController::class, 'update']);
-        Route::delete('{webinar}', [WebinarController::class, 'destroy']);
+        Route::delete('{webinar}', [WebinarController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{webinar:slug}', [WebinarController::class, 'destroy']);
     });
 
@@ -132,10 +162,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [PilierController::class, 'index']);
         Route::get('{pilier}', [PilierController::class, 'show']);
         Route::get('slug/{pilier:slug}', [PilierController::class, 'show']);
-        Route::post('/', [PilierController::class, 'store']);
-        Route::put('{pilier}', [PilierController::class, 'update']);
+        Route::post('/', [PilierController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{pilier}', [PilierController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{pilier:slug}', [PilierController::class, 'update']);
-        Route::delete('{pilier}', [PilierController::class, 'destroy']);
+        Route::delete('{pilier}', [PilierController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{pilier:slug}', [PilierController::class, 'destroy']);
     });
 
@@ -144,10 +174,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [PrivilegeController::class, 'index']);
         Route::get('{privilege}', [PrivilegeController::class, 'show']);
         Route::get('slug/{privilege:slug}', [PrivilegeController::class, 'show']);
-        Route::post('/', [PrivilegeController::class, 'store']);
-        Route::put('{privilege}', [PrivilegeController::class, 'update']);
+        Route::post('/', [PrivilegeController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{privilege}', [PrivilegeController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{privilege:slug}', [PrivilegeController::class, 'update']);
-        Route::delete('{privilege}', [PrivilegeController::class, 'destroy']);
+        Route::delete('{privilege}', [PrivilegeController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{privilege:slug}', [PrivilegeController::class, 'destroy']);
     });
 
@@ -156,10 +186,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [EvaFeatureController::class, 'index']);
         Route::get('{evaFeature}', [EvaFeatureController::class, 'show']);
         Route::get('slug/{evaFeature:slug}', [EvaFeatureController::class, 'show']);
-        Route::post('/', [EvaFeatureController::class, 'store']);
-        Route::put('{evaFeature}', [EvaFeatureController::class, 'update']);
+        Route::post('/', [EvaFeatureController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{evaFeature}', [EvaFeatureController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{evaFeature:slug}', [EvaFeatureController::class, 'update']);
-        Route::delete('{evaFeature}', [EvaFeatureController::class, 'destroy']);
+        Route::delete('{evaFeature}', [EvaFeatureController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{evaFeature:slug}', [EvaFeatureController::class, 'destroy']);
     });
 
@@ -168,10 +198,10 @@ Route::prefix('v1')->group(function () {
         Route::get('/', [OfferController::class, 'index']);
         Route::get('{offer}', [OfferController::class, 'show']);
         Route::get('slug/{offer:slug}', [OfferController::class, 'show']);
-        Route::post('/', [OfferController::class, 'store']);
-        Route::put('{offer}', [OfferController::class, 'update']);
+        Route::post('/', [OfferController::class, 'store'])->middleware(['auth:api', 'admin']);
+        Route::put('{offer}', [OfferController::class, 'update'])->middleware(['auth:api', 'admin']);
         Route::put('slug/{offer:slug}', [OfferController::class, 'update']);
-        Route::delete('{offer}', [OfferController::class, 'destroy']);
+        Route::delete('{offer}', [OfferController::class, 'destroy'])->middleware(['auth:api', 'admin']);
         Route::delete('slug/{offer:slug}', [OfferController::class, 'destroy']);
     });
